refactor : 전체 코드 MVC 패턴으로 리팩토링
