+++ src/main/java/calculator/Application.java
@@ -1,8 +1,10 @@
 package calculator;
 
+import calculator.controller.CalculatorController;
+
 public class Application {
     public static void main(String[] args) {
-        Calculator calculator = new Calculator();
-        calculator.start();
+        CalculatorController calculatorController = new CalculatorController();
+        calculatorController.run();
     }
 }
