+++ src/main/java/calculator/model/NumberParser.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.model;
 
 import java.util.ArrayList;
 import java.util.List;
