+++ src/main/java/calculator/model/StringAdd.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.model;
 
 import java.util.List;
