+++ src/main/java/calculator/model/Validator.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.model;
 
 public class Validator {
